Extract FileObjectSource, CacheRead and CacheWrite to cache_io.rs
Also, put get_file_mode and set_file_mode to utils.rs
