+++ src/cache/utils.rs
@@ -1,0 +1,44 @@
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use fs_err as fs;
+
+use std::path::Path;
+
+use crate::errors::*;
+
+#[cfg(unix)]
+pub(in crate::cache) fn get_file_mode(file: &fs::File) -> Result<Option<u32>> {
+    use std::os::unix::fs::MetadataExt;
+    Ok(Some(file.metadata()?.mode()))
+}
+
+#[cfg(windows)]
+#[allow(clippy::unnecessary_wraps)]
+pub(in crate::cache) fn get_file_mode(_file: &fs::File) -> Result<Option<u32>> {
+    Ok(None)
+}
+
+#[cfg(unix)]
+pub(in crate::cache) fn set_file_mode(path: &Path, mode: u32) -> Result<()> {
+    use std::fs::Permissions;
+    use std::os::unix::fs::PermissionsExt;
+    let p = Permissions::from_mode(mode);
+    fs::set_permissions(path, p)?;
+    Ok(())
+}
+
+#[cfg(windows)]
+#[allow(clippy::unnecessary_wraps)]
+pub(in crate::cache) fn set_file_mode(_path: &Path, _mode: u32) -> Result<()> {
+    Ok(())
+}
